interim refactor of rule groupings
