+++ respark/rules/generation_rules/generate_datetime.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from pyspark.sql import Column, functions as F
-from .core_rules import register_generation_rule, GenerationRule
+from ..rules_registry import register_generation_rule, GenerationRule
 from respark.random import randint_int, randint_long
