+++ respark/rules/generation_rules/generate_fractional.py
@@ -1,6 +1,6 @@
 from typing import Literal
 from pyspark.sql import Column, functions as F
-from .core_rules import register_generation_rule, GenerationRule
+from ..rules_registry import register_generation_rule, GenerationRule
 from respark.core import FRACTIONAL_BOUNDS, FRACTIONAL_CAST
