+++ respark/rules/generation_rules/generate_integral.py
@@ -1,6 +1,6 @@
 from typing import Literal
 from pyspark.sql import Column
-from .core_rules import register_generation_rule, GenerationRule
+from ..rules_registry import register_generation_rule, GenerationRule
 from respark.core import INTEGRAL_BOUNDS, INTEGRAL_CAST
 from respark.random import randint_int
