+++ respark/rules/generation_rules/generate_string.py
@@ -1,6 +1,6 @@
 import string
 from pyspark.sql import Column, functions as F, types as T
-from .core_rules import register_generation_rule, GenerationRule
+from ..rules_registry import register_generation_rule, GenerationRule
 from respark.random import randint_int, choice
